+++ ui/src/commonMain/kotlin/tech/kotlinlang/ui/AmountCurrency.kt
@@ -1,0 +1,8 @@
+package tech.kotlinlang.ui
+
+sealed class AmountCurrency {
+    data object IndianRupee: AmountCurrency()
+
+    val symbol: String
+        get() = "₹"
+}